Extract mDNSListener from base.go
